Resolved properties to work with spring boot
