--- src/main/java/org/owasp/dependencytrack/config/DatabaseConfguration.java
@@ -1,69 +1,0 @@
-package org.owasp.dependencytrack.config;
-
-import org.apache.commons.dbcp.BasicDataSource;
-import org.owasp.dependencytrack.model.Application;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.orm.hibernate4.LocalSessionFactoryBean;
-
-import javax.sql.DataSource;
-import java.util.Properties;
-
-/**
- * Created by jason on 16/11/15.
- */
-@Configuration
-@Import(PropertyConfiguration.class)
-public class DatabaseConfguration
-{
-    @Value("${jdbc.driverClassName}")
-    private String driverClassName;
-
-    @Value("${jdbc.username}")
-    private String username;
-
-    @Value("${jdbc.password}")
-    private String password;
-
-    @Value("${jdbc.url}")
-    private String url;
-
-    @Value("${hibernate.dialect}")
-    private String hibernateDialect;
-
-    @Value("${hibernate.show_sql}")
-    private String hibernateShowSql;
-
-    @Value("${hibernate.hbm2ddl.auto}")
-    private String hibernateHbm2ddlAuto;
-
-    @Bean
-    public DataSource dataSource(){
-        BasicDataSource basicDataSource = new BasicDataSource();
-        basicDataSource.setDriverClassName(driverClassName);
-        basicDataSource.setUsername(username);
-        basicDataSource.setPassword(password);
-        basicDataSource.setUrl(url);
-        return basicDataSource;
-    }
-
-    @Bean
-    public LocalSessionFactoryBean sessionFactory(){
-        LocalSessionFactoryBean localSessionFactoryBean = new LocalSessionFactoryBean();
-        localSessionFactoryBean.setDataSource(dataSource());
-        localSessionFactoryBean.setHibernateProperties(hibernateProperties());
-        localSessionFactoryBean.setPackagesToScan(Application.class.getPackage().getName());
-        return localSessionFactoryBean;
-    }
-
-    @Bean
-    public Properties hibernateProperties(){
-        Properties properties = new Properties();
-        properties.setProperty("hibernate.dialect", hibernateDialect);
-        properties.setProperty("hibernate.show_sql", hibernateShowSql);
-        properties.setProperty("hibernate.hbm2ddl.auto", hibernateHbm2ddlAuto);
-        return properties;
-    }
-}
