+++ src/main/java/org/owasp/dependencytrack/config/DatabaseConfiguration.java
@@ -1,0 +1,46 @@
+package org.owasp.dependencytrack.config;
+
+import org.apache.commons.dbcp.BasicDataSource;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.orm.jpa.vendor.HibernateJpaSessionFactoryBean;
+
+import javax.sql.DataSource;
+
+/**
+ * Created by jason on 16/11/15.
+ */
+@Configuration
+@Import(PropertyConfiguration.class)
+public class DatabaseConfiguration
+{
+    @Value("${spring.datasource.driverClassName}")
+    private String driverClassName;
+
+    @Value("${spring.datasource.username}")
+    private String username;
+
+    @Value("${spring.datasource.password}")
+    private String password;
+
+    @Value("${spring.datasource.url}")
+    private String url;
+
+    @Bean
+    public DataSource dataSource(){
+        BasicDataSource basicDataSource = new BasicDataSource();
+        basicDataSource.setDriverClassName(driverClassName);
+        basicDataSource.setUsername(username);
+        basicDataSource.setPassword(password);
+        basicDataSource.setUrl(url);
+        return basicDataSource;
+    }
+
+    @Bean
+    public HibernateJpaSessionFactoryBean sessionFactory() {
+        return new HibernateJpaSessionFactoryBean();
+    }
+
+}
